+++ src/plugins/index.ts
@@ -5,6 +5,7 @@ import { alignmentPlugin } from './alignment';
 import { linksPlugin } from './links';
 import { videoPlugin } from './video';
 import { imagePlugin } from './image';
+import { colorPlugin } from './color';
 
 export { basicFormattingPlugin } from './basicFormatting';
 export { headingsPlugin } from './headings';
@@ -13,11 +14,13 @@ export { alignmentPlugin } from './alignment';
 export { linksPlugin } from './links';
 export { videoPlugin } from './video';
 export { imagePlugin } from './image';
+export { colorPlugin } from './color';
 
 // Export modal wrappers
 export { LinkModalWrapper } from './links';
 export { VideoModalWrapper } from './video';
 export { ImageModalWrapper } from './image';
+export { ColorButton } from './color';
 
 // Export a default set of plugins
 export const defaultPlugins = [
@@ -27,6 +30,7 @@ export const defaultPlugins = [
   alignmentPlugin,
   linksPlugin,
   videoPlugin,
-  imagePlugin
+  imagePlugin,
+  colorPlugin
 ];